Inline NeMo Gym helpers
